fix: update quantity handling to use float instead of int for precision

+++ electron/domain/budget.js
@@ -90,7 +90,7 @@ class Budget {
     if (!this.hasProducts()) return 0;
     
     return this.budgetProducts.reduce((sum, product) => {
-      return sum + (parseFloat(product.unit_price) * parseInt(product.quantity));
+      return sum + (parseFloat(product.unit_price) * parseFloat(product.quantity));
     }, 0);
   }
