+++ electron/services/budgetService.js
@@ -164,7 +164,7 @@ class BudgetService {
         items: items.map(item => ({
           product_id: item.product_id ? parseInt(item.product_id) : null,
           template_id: item.template_id ? parseInt(item.template_id) : null,
-          quantity: parseInt(item.quantity),
+          quantity: parseFloat(item.quantity),
           unit_price: parseFloat(item.unit_price)
         }))
       }
